see #104 Fixed all issues in Swagger Editor and issue 104. (#106)
**Swagger Editor Issues:**
1. When I opened the file in Swagger Editor, I noticed three errors in
the inspector.
- The first error was resolved by changing the `openapi` version from
`3.1.0` to `3.0.0` or `3.1.1`. If you want to keep version `3.1.0`, it
needs to be updated to `3.1.1`. According to the official OpenAPI
documentation, there's no difference between the two versions.
- The second error was a semantic issue: **GET operations cannot have a
requestBody.** I fixed this by removing the requestBody in line 1688.
- The third issue was a structural error: **additionalProperty: ref$**.
I corrected this by changing `ref$` to `$ref` in line 14355.

**Issue 104:**
In OpenAPI, external URLs should be approved for reference. To meet this
requirement, I added `https://raw.githubusercontent.com` to the servers
section of the specification at line 25.